requirements.txt
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
fastapi==0.78.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
requests==2.27.1
Rx==3.2.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tcd==3.2.2
twitch-python==0.0.20
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
